+++ digest/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.11.4 (UNRELEASED)
+### Added
+- Optional `impl:` section in the `buffer_ct_variable!` macro with support for `CustomizedInit` ([#NNNN])
+
+[#NNNN]: https://github.com/RustCrypto/traits/pull/NNNN
+
 ## 0.11.3 (2026-04-03)
 ### Added
 - `dev::initialized_mac_test` function ([#2367])